tests/wpt/web-platform-tests/portals/about-blank-cannot-host.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

promise_test(async (t) => {
  let hostWindow = window.open();
  assert_equals(hostWindow.location.href, "about:blank");

  let portal = hostWindow.document.createElement("portal");
  portal.src = "resources/simple-portal.html";
  hostWindow.document.body.appendChild(portal);

  await promise_rejects(t, "InvalidStateError", portal.activate());
}, "about:blank cannot host a portal");

</script>
